chore: re-organize the folder structure
