+++ npc/FutureCore/src/main/scala/futurecore/Elaborate.scala
@@ -17,20 +17,16 @@
   external case of the White Rabbit switch or other product you make using
   this documentation.
  */
-package futurecore.frontend
+import circt.stage._
 
-import chisel3._
-
-class ProgramCounterIO extends Bundle {
-  val instAddr = Output(UInt(32.W))
-}
-
-class ProgramCounter extends Module {
-  val io        = IO(new ProgramCounterIO)
-  val instWidth = 32.W
-
-  val pc = RegInit("h80000000".U(instWidth)) // PC reset to 0x80000000
-  pc := pc + 4.U
-
-  io.instAddr := pc
+object Elaborate extends App {
+  def top       = new FutureCore
+  val generator = Seq(chisel3.stage.ChiselGeneratorAnnotation(() => top))
+  (new ChiselStage)
+    .execute(
+      args,
+      generator
+        :+ CIRCTTargetAnnotation(CIRCTTarget.SystemVerilog)
+        // :+ FirtoolOption("-O=debug")
+    )
 }
