+++ npc/Makefile
@@ -1,17 +1,15 @@
 export PATH := $(PATH):$(abspath ./utils)
 
-TOP_NAME = FutureCore
-PRJ_DIR = $(shell echo $(TOP_NAME) | tr A-Z a-z)
-SRC_DIR = $(PRJ_DIR)/src
-SIM_DIR = $(PRJ_DIR)/sim
+PRJ_NAME = FutureCore
+SRC_DIR = $(PRJ_NAME)/src
+SIM_DIR = $(PRJ_NAME)/sim
 
 test:
 	mill -i __.test
 
 verilog:
 	$(call git_commit, "generate verilog")
-	mkdir -p $(SIM_DIR)/vsrc
-	mill -i __.runMain Elaborate -td $(SIM_DIR)/vsrc
+	mill -i __.runMain Elaborate -td $(SIM_DIR)
 
 help:
 	mill -i __.runMain Elaborate --help
@@ -29,7 +27,7 @@ checkformat:
 	mill -i __.checkFormat
 
 clean:
-	-rm $(SIM_DIR)/vsrc/$(TOP_NAME).sv
+	-rm $(SIM_DIR)/$(TOP_NAME).sv
 	make -C $(SIM_DIR) clean
 
 sim:
